chroe: rename the folder
