+++ packages/error-overlay/src/index.ts
@@ -1,4 +1,4 @@
 // This package 'error-overlay' is a modified version of the Next.js that can be found here:
 // https://github.com/vercel/next.js/blob/canary/packages/react-dev-overlay/
 
-export { default as ReactDevOverlay } from './react/ReactDevOverlay';
+export { default as ErrorOverlay } from './view/ErrorOverlay';
