+++ packages/error-overlay/src/view/components/Dialog/Dialog.tsx
