+++ packages/error-overlay/src/view/components/Dialog/DialogBody.tsx
