+++ packages/error-overlay/src/view/components/Dialog/DialogContent.tsx
