+++ packages/error-overlay/src/view/components/Dialog/DialogHeader.tsx
