+++ packages/error-overlay/src/view/components/Dialog/index.ts
