+++ packages/error-overlay/src/view/components/Dialog/styles.ts
