+++ packages/error-overlay/src/view/components/Errors/CodeFrame/CodeFrame.tsx
