+++ packages/error-overlay/src/view/components/Errors/styles.ts
