+++ packages/error-overlay/src/view/components/Icons/ArrowLeftIcon.tsx
