+++ packages/error-overlay/src/view/components/Icons/ArrowRightIcon.tsx
