+++ packages/error-overlay/src/view/components/Icons/CloseIcon.tsx
