+++ packages/error-overlay/src/view/components/Icons/ExternalLinkIcon.tsx
