+++ packages/error-overlay/src/view/components/Icons/WarningIcon.tsx
