+++ packages/error-overlay/src/view/components/Icons/index.ts
