+++ packages/error-overlay/src/view/components/NavigationBar/NavigationBar.tsx
