+++ packages/error-overlay/src/view/components/NavigationBar/index.ts
