+++ packages/error-overlay/src/view/components/NavigationBar/styles.ts
