+++ packages/error-overlay/src/view/components/Overlay/body-locker.ts
