+++ packages/error-overlay/src/view/components/Overlay/styles.tsx
