+++ packages/error-overlay/src/view/components/Terminal/Terminal.tsx
