+++ packages/error-overlay/src/view/components/Terminal/index.tsx
