+++ packages/error-overlay/src/view/components/Terminal/styles.tsx
