+++ packages/error-overlay/src/view/errorTypeHandler.ts
