+++ packages/error-overlay/src/view/helpers/getErrorByType.ts
