+++ packages/error-overlay/src/view/helpers/noop-template.ts
