+++ packages/error-overlay/src/view/helpers/stack-frame.ts
