+++ packages/error-overlay/src/view/hooks/use-on-click-outside.ts
