+++ packages/error-overlay/src/view/styles/Base.tsx
