+++ packages/error-overlay/src/view/styles/ComponentStyles.tsx
