+++ packages/error-overlay/src/view/styles/index.tsx
